# ogrinfo: speed-up on GPKG with large number of layers