--- a/apps/ogrinfo_lib.cpp
+++ b/apps/ogrinfo_lib.cpp
@@ -129,6 +129,14 @@ static void Concat(CPLString &osRet, bool bStdoutOutput, const char *pszFormat,
     va_end(args);
 }
 
+static void ConcatStr(CPLString &osRet, bool bStdoutOutput, const char *pszStr)
+{
+    if (bStdoutOutput)
+        fwrite(pszStr, 1, strlen(pszStr), stdout);
+    else
+        osRet += pszStr;
+}
+
 /************************************************************************/
 /*                        ReportFieldDomain()                           */
 /************************************************************************/
@@ -515,10 +523,9 @@ static void ReportRelationships(CPLString &osRet, CPLJSONObject &oRoot,
                 for (const auto &osName : aosList)
                 {
                     if (!bFirstName)
-                        Concat(osRet, psOptions->bStdoutOutput, ", ");
+                        ConcatStr(osRet, psOptions->bStdoutOutput, ", ");
                     bFirstName = false;
-                    Concat(osRet, psOptions->bStdoutOutput, "%s",
-                           osName.c_str());
+                    ConcatStr(osRet, psOptions->bStdoutOutput, osName.c_str());
                 }
                 Concat(osRet, psOptions->bStdoutOutput, "\n");
             };
@@ -1051,8 +1058,8 @@ static void ReportOnLayer(CPLString &osRet, CPLJSONObject oLayer,
                     }
                     else
                     {
-                        Concat(osRet, psOptions->bStdoutOutput, "%s",
-                               poSupportedSRS->GetName());
+                        ConcatStr(osRet, psOptions->bStdoutOutput,
+                                  poSupportedSRS->GetName());
                     }
                 }
                 Concat(osRet, psOptions->bStdoutOutput, "\n");
@@ -1375,8 +1382,8 @@ static void ReportOnLayer(CPLString &osRet, CPLJSONObject oLayer,
                 }
                 else
                 {
-                    Concat(
-                        osRet, psOptions->bStdoutOutput, "%s",
+                    ConcatStr(
+                        osRet, psOptions->bStdoutOutput,
                         poFeature
                             ->DumpReadableAsString(psOptions->aosOptions.List())
                             .c_str());
@@ -1396,9 +1403,10 @@ static void ReportOnLayer(CPLString &osRet, CPLJSONObject oLayer,
         }
         else
         {
-            Concat(osRet, psOptions->bStdoutOutput, "%s",
-                   poFeature->DumpReadableAsString(psOptions->aosOptions.List())
-                       .c_str());
+            ConcatStr(
+                osRet, psOptions->bStdoutOutput,
+                poFeature->DumpReadableAsString(psOptions->aosOptions.List())
+                    .c_str());
             OGRFeature::DestroyFeature(poFeature);
         }
     }
@@ -1416,7 +1424,7 @@ static void PrintLayerSummary(CPLString &osRet, CPLJSONObject &oLayer,
     if (bJson)
         oLayer.Set("name", poLayer->GetName());
     else
-        Concat(osRet, psOptions->bStdoutOutput, "%s", poLayer->GetName());
+        ConcatStr(osRet, psOptions->bStdoutOutput, poLayer->GetName());
 
     const char *pszTitle = poLayer->GetMetadataItem("TITLE");
     if (pszTitle)
@@ -1448,8 +1456,8 @@ static void PrintLayerSummary(CPLString &osRet, CPLJSONObject &oLayer,
             {
                 if (iGeom > 0)
                     Concat(osRet, psOptions->bStdoutOutput, ", ");
-                Concat(osRet, psOptions->bStdoutOutput, "%s",
-                       OGRGeometryTypeToName(poGFldDefn->GetType()));
+                ConcatStr(osRet, psOptions->bStdoutOutput,
+                          OGRGeometryTypeToName(poGFldDefn->GetType()));
             }
         }
         if (!bJson)
@@ -1886,14 +1894,15 @@ char *GDALVectorInfo(GDALDatasetH hDataset,
     if (bJson)
     {
         osRet.clear();
-        Concat(osRet, psOptions->bStdoutOutput, "%s",
-               json_object_to_json_string_ext(
-                   static_cast<struct json_object *>(oRoot.GetInternalHandle()),
-                   JSON_C_TO_STRING_PRETTY
+        ConcatStr(
+            osRet, psOptions->bStdoutOutput,
+            json_object_to_json_string_ext(
+                static_cast<struct json_object *>(oRoot.GetInternalHandle()),
+                JSON_C_TO_STRING_PRETTY
 #ifdef JSON_C_TO_STRING_NOSLASHESCAPE
-                       | JSON_C_TO_STRING_NOSLASHESCAPE
+                    | JSON_C_TO_STRING_NOSLASHESCAPE
 #endif
-                   ));
+                ));
     }
 
     return VSI_STRDUP_VERBOSE(osRet);