On the dataset of https://github.com/qgis/QGIS/issues/53525, this speeds up ogrinfo -json by a factor of 2.

Before:
```
$ time ogrinfo -json file_300_filled_layers.gpkg > out.json
real	0m0,565s
user	0m0,524s
sys	0m0,040s
```

After:
```
$ time ogrinfo -json file_300_filled_layers.gpkg > out.json
real	0m0,258s
user	0m0,217s
sys	0m0,041s
```